requirements.txt
pydantic>=2.0.0
requests>=2.28.0
python-dateutil>=2.8.0
typing-extensions>=4.0.0
pytest>=7.0.0
openai>=2.8.1
python-dotenv>=1.0.0
PyYAML>=6.0.0
ddgs>=3.3.0
beautifulsoup4>=4.12.0
aiohttp>=3.8.0
